using populate to revise the userProfile controller. And Bug fix(for hashing password, getEdit&postEdit&deleteVideo)

+++ src/controllers/videoController.js
@@ -18,19 +18,27 @@ export const watch = async(req, res) => {
 
 export const getEdit = async(req, res) => {
     const { id } = req.params;
+    const { user: {_id} } = req.session
     const video = await Video.findById(id);
     if (!video) {
         return res.status(404).render("404", {pageTitle: `Error: Video not found`})
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle: `Edit: ${video.title}`, video })
 };
 
 export const postEdit = async(req, res) => {
     const { id } = req.params;
+    const { user: {_id} } = req.session
     const video = await Video.exists({ _id: id });
     const { title, description, hashtags } = req.body
     if (!video) {
         return res.status(404).render("404", {pageTitle: `Error: Video not found`})
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -58,7 +66,7 @@ export const postUpload = async(req, res) => {
         }
     } = req;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
             fileUrl:file.path,
@@ -70,6 +78,9 @@ export const postUpload = async(req, res) => {
                 rating : 0
             }
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save()
         return res.redirect("/");
     } catch (error) {
         return res.status(400).render("upload", {
@@ -81,6 +92,14 @@ export const postUpload = async(req, res) => {
 
 export const deleteVideo = async(req, res) => {
     const { id } = req.params;
+    const { user: {_id} } = req.session
+    const video = await VideolfindBYId(id);
+    if (!video) {
+        return res.status(404).render("404", {pageTitle: `Error: Video not found`})
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/")
 }
